### 2023-11-05 #5998 @arfon
@LuisScoccola @annoviko @tahirinadia @amarkos – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _CM++ - A Meta-method for Well-Connected Community Detection_

Based on your experience, (and for some of you, past submissions to JOSS) we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
