+++ README.md
@@ -1,5 +1,7 @@
 # CM++ Pipeline
 
+[![DOI](https://zenodo.org/badge/599799149.svg)](https://zenodo.org/doi/10.5281/zenodo.10076513)
+
 Customizable modular pipeline for testing an improved version of CM for generating well-connected clusters. Image below from arXiv preprint: Park et. al. (2023). https://github.com/illinois-or-research-analytics/cm_pipeline/tree/main
 
 - [CM++ Pipeline](#cm-pipeline)
