### 2023-11-21 #6073 @editorialbot
[REVIEW]: CM++ - A Meta-method for Well-Connected Community Detection
**Submitting author:** @chackoge (<a href="http://orcid.org/0000-0002-2127-1892">George Chacko</a>)
**Repository:** https://github.com/illinois-or-research-analytics/cm_pipeline.git
**Branch with paper.md** (empty if default branch): main
**Version:** v4.0.0
