+++ requirements.txt
@@ -1,35 +1,33 @@
-attrs==22.2.0
-click==8.1.3
-colorama==0.4.6
-coloredlogs==15.0.1
-exceptiongroup==1.1.1
-graphviz==0.20.1
-HeapDict==1.0.1
+attrs==22.2
+click==8.1
+colorama==0.4
+coloredlogs==15.0
+exceptiongroup==1.1
+graphviz==0.20
+HeapDict==1.0
 humanfriendly==10.0
-igraph==0.10.4
-iniconfig==2.0.0
-jsonpickle==2.2.0
-leidenalg==0.9.1
+igraph==0.10
+iniconfig==2.0
+jsonpickle==2.2
+leidenalg==0.9
 networkit==10.1
-numpy==1.24.2
+numpy==1.24
 packaging==23.0
-pandas==1.5.3
+pandas==1.5
 pip
-pluggy==1.0.0
-pytest==7.2.2
-python-dateutil==2.8.2
+pluggy==1.0
+pytest==7.2
 pytz==2023.2
-scipy==1.10.1
-setuptools==50.3.2
-six==1.16.0
-structlog==22.3.0
-texttable==1.6.7
-tomli==2.0.1
-treeswift==1.1.33
-typer==0.6.1
-typing-extensions==4.5.0
+scipy==1.10
+setuptools==50.3
+six==1.16
+structlog==22.3
+tomli==2.0
+treeswift==1.1
+typer==0.6
+typing-extensions==4.5
 networkx==3.1
-psutil==5.9.5
-infomap==2.7.1
+psutil==5.9
+infomap==2.7
 git+https://github.com/vikramr2/python-mincut
 .
