+++ README.md
@@ -10,6 +10,9 @@ Customizable modular pipeline for testing an improved version of CM for generati
     - [Setup via Cloning](#setup-via-cloning)
     - [Setup via pip install](#setup-via-pip-install)
   - [Input and Usage](#input-and-usage)
+    - [Example Commands](#example-commands)
+      - [CM++](#cm)
+      - [CM Pipeline](#cm-pipeline-1)
     - [CM++ Usage](#cm-usage)
     - [Pipeline Usage](#pipeline-usage)
       - [JSON Input Documentation](#json-input-documentation)
@@ -43,6 +46,21 @@ Customizable modular pipeline for testing an improved version of CM for generati
 Simply run `pip install git+https://github.com/illinois-or-research-analytics/cm_pipeline`. This will install CM++, but to use pipeline functionality, please setup via cloning.
 
 ## Input and Usage
+
+### Example Commands
+
+#### CM++
+
+- `python3 -m hm01.cm -i network.tsv -e clustering.tsv -o output.tsv -c leiden -g 0.5 --threshold 1log10 --nprocs 4 --quiet`
+  - Runs CM++ on a Leiden with resolution 0.5 clustering with connectivity threshold $log_{10}(n)$ (Every cluster with connectivity over the log of the number of nodes is considered "well-connected")
+- `python3 -m hm01.cm -i network.tsv -e clustering.tsv -o output.tsv -c ikc -k 10 --threshold 1log10 --nprocs 4 --quiet`
+  - Similar idea but with IKC having hyperparameter $k=10$.
+
+#### CM Pipeline
+
+- Suppose you have a pipeline like the one [here](examples/leiden.json). Call it `pipeline.json`
+- Then from the root of this repository run:
+  - `python -m main pipeline.json`
 
 ### CM++ Usage
 
