+++ docs/dev_setup.md
@@ -1,0 +1,14 @@
+# For Developers
+
+## Loading a Developer Environment
+
+To quickly set up a developer environment for the CM++ Pipeline, simply run the following commands. (**NOTE: Make sure you have Conda installed**)
+
+```bash
+conda env create -f environment.yml
+conda activate 
+```
+
+## Customizing the Pipeline
+
+- The CM++ Pipeline also allows for users to add their own pipeline stages and clustering methods.
