+++ docs/features.md
@@ -1,0 +1,5 @@
+# Features
+
+The CM Pipeline is a modular pipeline for community detection that contains the following modules:
+
+- **Cluster Statistics**: Compute statistics such as node and edge count,
