+++ docs/index.md
@@ -2,86 +2,25 @@
 
 [![DOI](https://zenodo.org/badge/599799149.svg)](https://zenodo.org/doi/10.5281/zenodo.10076513)
 
-Customizable modular pipeline for testing an improved version of CM for generating well-connected clusters. Image below from arXiv preprint: Park et. al. (2023). https://github.com/illinois-or-research-analytics/cm_pipeline/tree/main
+Customizable modular pipeline for testing an improved version of CM for generating well-connected clusters. Image below from arXiv preprint: Park et. al. (2023). [(Original Source Code)](https://github.com/illinois-or-research-analytics/cm_pipeline/tree/main)
 
 ![cm_pipeline Overview](figures/cm_pp_overview.png)
 
 ## Quick Start Guide
 
-### Features
-
-The CM Pipeline is a modular pipeline for community detection that contains the following modules:
-
-- **Cluster Statistics**: Compute statistics such as node and edge count, 
-
-### Requirements
-
-- MacOS or Linux operating system
-- `python3.9` or higher
-- `cmake 3.2.0` or higher
-- `gcc` of any version (In our analysis, `gcc 9.2.0` was used)
-
-### Installation and Setup
-
-There are several strategies for installation
-
-#### Installation via Cloning
-
-- Clone the cm_pipeline repository
-- Activate the venv which has the necessary packages
-- Run `pip install -r requirements.txt && pip install .`
-- Make sure everything installed properly by running `cd tests && pytest`
-
-#### Installation via pip install
-
-Simply run `pip install git+https://github.com/illinois-or-research-analytics/cm_pipeline`. **This will install CM++, but to use pipeline functionality, please setup via cloning.**
+- [Features](features.md)
+- [Installation](installation.md)
+- [Quick Start Guide](quickstart.md)
 
 ## Input and Usage
 
-### Example Commands
+- [Example Commands](examples.md)
+- [CM++ Usage](cmpp_usage.md)
+- [CM Pipeline Usage](pipeline_usage.md)
 
-#### CM++
+## Contributing
 
-- `python3 -m hm01.cm -i network.tsv -e clustering.tsv -o output.tsv -c leiden -g 0.5 --threshold 1log10 --nprocs 4 --quiet`
-  - Runs CM++ on a Leiden with resolution 0.5 clustering with connectivity threshold $log_{10}(n)$ (Every cluster with connectivity over the log of the number of nodes is considered "well-connected")
-- `python3 -m hm01.cm -i network.tsv -e clustering.tsv -o output.tsv -c ikc -k 10 --threshold 1log10 --nprocs 4 --quiet`
-  - Similar idea but with IKC having hyperparameter $k=10$.
-
-#### CM Pipeline
-
-- Then from the root of this repository run:
-  - `python -m main pipeline.json`
-
-### CM++ Usage
-
-### Pipeline Usage
-
-- Edit the fields of the `pipeline.json` file to reflect your inputs and requirements.
-- Run `python -m main pipeline.json`
-
-#### JSON Input Documentation
-
-
-## For Developers
-
-### Loading a Developer Environment
-
-To quickly set up a developer environment for the CM++ Pipeline, simply run the following commands. (**NOTE: Make sure you have Conda installed**)
-
-```bash
-conda env create -f environment.yml
-conda activate 
-```
-
-### Customizing the Pipeline
-
-- The CM++ Pipeline also allows for users to add their own pipeline stages and clustering methods.
-
-## Output Files
-
-- The commands executed during the workflow are captured in `{output_dir}/{run_name}-{timestamp}/commands.sh`. This is the shell script generated by the pipeline that is run to generate outputs.
[…]
