### 2023-12-07 0a7b6e0
mathjax is wonky wih markdown
