+++ docs/index.rst
@@ -1,0 +1,1 @@
+.. include:: index.md
