+++ docs/conf.py
@@ -1,3 +1,4 @@
+# pylint: disable=W0622
 # Configuration file for the Sphinx documentation builder.
 #
 # For the full list of built-in configuration values, see the documentation:
@@ -32,10 +33,7 @@ html_static_path = ['_static']
 extensions = ['myst_parser']
 
 master_doc = 'index'
-source_suffix = {
-    '.rst': 'restructuredtext',
-    '.md': 'markdown',
-}
+source_suffix = '.rst'
 
 
 html_baseurl = 'cm_pipeline'
