+++ docs/index.rst
@@ -1,5 +1,0 @@
-=============
-CM++ Pipeline
-=============
-
-Customizable modular pipeline for testing an improved version of CM for generating well-connected clusters. Image below from arXiv preprint: Park et. al. (2023). `(Original Source Code) <https://github.com/illinois-or-research-analytics/cm_pipeline/tree/main>`__
