### 2023-12-07 6af2afb
trying mkdocs instead
