+++ doc/docs/cmpp_usage.md
@@ -1,1 +1,0 @@
-# CM++ Usage
