+++ doc/docs/quickstart.md
@@ -1,1 +1,0 @@
-# Quick Start Guide
