+++ doc/docs/pipeline_usage.md
@@ -5,6 +5,8 @@
 
 ## JSON Input Documentation
 
+To create a JSON pipeline that fits your needs, see the following [tutorial](json_format.md)
+
 ## Output Files
 
 - The commands executed during the workflow are captured in `{output_dir}/{run_name}-{timestamp}/commands.sh`. This is the shell script generated by the pipeline that is run to generate outputs.
