+++ doc/docs/dev_setup.md
@@ -12,3 +12,5 @@ conda activate
 ## Customizing the Pipeline
 
 - The CM++ Pipeline also allows for users to add their own pipeline stages and clustering methods.
+- See the following tutorial for adding new stages and algorithms to the pipeline:
+    - [Pipeline Customization](pipeline_customization.md)
