+++ doc/docs/quickstart.md
@@ -1,0 +1,8 @@
+# Quick Start Guide
+
+Suppose we have a network that we want to cluster and refine using the pipeline. For this tutorial, I will use the following network and clustering.
+
+- [network.tsv]()
+- [clustering.tsv]()
+
+First, let's visualize the graph and clustering
