+++ doc/docs/contributions.md
@@ -1,0 +1,42 @@
+# Contributing to the CM++ Pipeline
+
+We welcome all helpful contributions to improving CM++ and the CM pipeline. Contributors can help in the following ways:
+
+- Identifying and/or fixing bugs and typos
+- Writing tutorials and blog posts
+- Creating tests and adding them to the [test set](https://github.com/illinois-or-research-analytics/cm_pipeline/tree/main/tests)
+- Proposing new features
+
+## How to Contribute
+
+- Raise [issues](https://github.com/illinois-or-research-analytics/cm_pipeline/issues). Issues can include bugs and errors with installation and usage, or just suggestions for new features. Before raising an issue, please make sure of the following:
+    - Double check that your issue isnt already addressed in the documentation on this website.
+    - If you are running into problems with code, ensure that the dependencies and your version of the CM++ Pipeline are up to date
+    - Check if an issue has already been posted that addresses your concern.
+- Commit and pull request to a forked repository
+- Adding tutorials and blog posts to this wiki
+
+## Community Guidelines
+
+- Please make contributions on a forked repository and we will approve pull requests
+- CM++ is majority written in Python. Please ensure you are following the [style guidelines](https://peps.python.org/pep-0008/)
+- Please do not use submodules in your pull requests, package them and include them as a part of [requirements.txt](https://github.com/illinois-or-research-analytics/cm_pipeline/blob/main/requirements.txt)
+- Use appropiate version tags. We follow a standard `version.feature.patch` format for version tags.
+- Make sure you pass all test cases before making a pull request.
+
+### Using Appropriate Language
+
+In commit messages, wiki and blog-posts, and issues, we expect that contributors create a positive environment by:
+
+- Using welcoming and inclusive language
+- Being respectful of differing viewpoints and experiences
+- Focusing on what is best for the community
+- Showing empathy towards other community members
+
+We expect that contributors refrain from
+
+- The use of sexualized language or imagery and unwelcome sexual attention or advances
+- Insulting/derogatory comments, and personal or political attacks
+- Public or private harassment
+- Publishing your own or others’ private information, such as a physical or electronic addresses
+- Other conduct which could reasonably be considered inappropriate in a professional setting
