+++ doc/docs/quickstart.md
@@ -1,8 +1,203 @@
 # Quick Start Guide
 
-Suppose we have a network that we want to cluster and refine using the pipeline. For this tutorial, I will use the following network and clustering.
+## Running CM++
 
-- [network.tsv]()
-- [clustering.tsv]()
+Suppose you have a network and a clustering of the network that you want to refine. For this tutorial, I will use the following network and clustering.
 
-First, let's visualize the graph and clustering
+- [network.tsv](examples/network.tsv)
+- [clustering.tsv](examples/clustering.tsv)
+
+You know that these clusters aren't well-connected, and you want to use CM++ to enforce better connectivity.
+
+### Visualizing the Data
+
+First, you can visualize the graph and clustering. This can be done via the following Python script:
+
+```python
+import matplotlib.pyplot as plt
+import networkx as nx
+import pandas as pd
+
+# Load the clustering data
+clustering_file = 'clustering.tsv'
+clustering_data = pd.read_csv(clustering_file, sep='\t', header=None, names=['node_id', 'cluster_id'])
+
+# Load the edge list data
+edge_list_file = 'network.tsv'
+G = nx.read_edgelist(edge_list_file, delimiter='\t', nodetype=int)
+
+# Assign cluster IDs to nodes in the graph
+node_cluster_mapping = dict(clustering_data.values)
+nx.set_node_attributes(G, node_cluster_mapping, 'cluster')
+
+# Generate a color map for clusters
+unique_clusters = set(node_cluster_mapping.values())
+color_map = plt.cm.get_cmap('viridis', len(unique_clusters))
+colors = {cluster: color_map(i) for i, cluster in enumerate(unique_clusters)}
+
+# Assign color to each node based on cluster
+node_colors = [colors[node_cluster_mapping[node]] for node in G.nodes()]
+
+# Plot the graph
+plt.figure(figsize=(12, 8))
+nx.draw(G, node_color=node_colors, with_labels=False, node_size=50, edge_color='gray')
+plt.savefig('patitions.png')
+plt.show()
+```
+
+Running this script results in the following:
+
+![network](examples/patitions.png)
+
+### Installing and Running CM++
+
+CM++ can form these into well-connected clusters. Firstly, CM++ can be quickly installed with the following command:
+
+`pip install git+https://github.com/illinois-or-research-analytics/cm_pipeline`
+
+If you want all clusters to have a connectivity above `log10(n)` (all clusters have a mincut value above log10 of the number of nodes in the cluster), you can run the command below:
+
+`python3 -m hm01.cm -i network.tsv -e clustering.tsv -o out.tsv -c leiden -g 0.5 -t 1log10`
+
+### Visualize Post-CM Clustering
+
+CM++ pruned some nodes from this clustering to form well-connected clusters. You can visualize the clustering through the following script:
+
+```python
+import matplotlib.pyplot as plt
+import networkx as nx
+import pandas as pd
+
+# Load the clustering data
+clustering_file = 'out.tsv'
+clustering_data = pd.read_csv(clustering_file, sep='\t', header=None, names=['node_id', 'cluster_id'])
+
+# Load the edge list data
+edge_list_file = 'network.tsv'
+G = nx.read_edgelist(edge_list_file, delimiter='\t', nodetype=int)
+
+# Assign cluster IDs to nodes in the graph
+node_cluster_mapping = dict(clustering_data.values)
+nx.set_node_attributes(G, node_cluster_mapping, 'cluster')
+
+# Generate a color map for clusters
+unique_clusters = set(node_cluster_mapping.values())
+color_map = plt.cm.get_cmap('viridis', len(unique_clusters))
[…]
