### 2023-12-28 #6073 @LuisScoccola
Hi @chackoge! I just submitted an issue, probably related to the fact that I am running MacOS. Please let me know if you have any suggestions. I don't see the pdf response that you referred to in your last message, where can I find it?
