### 2023-12-28 #6073 @chackoge
Dear @LuisScoccola <https://github.com/LuisScoccola> thanks very much for
this feedback. We're working through your very helpful comments and hope to
have a response back in less than one week. We don't think it's a MacOS
issue- more about specifying dependencies but we'll report back. Happy New
Year to all of you.
George Chacko

On Thu, Dec 28, 2023 at 8:18 AM LuisScoccola ***@***.***>
wrote:

> Hi @chackoge <https://github.com/chackoge>! I just submitted an issue,
> probably related to the fact that I am running MacOS. Please let me know if
> you have any suggestions. I don't see the pdf response that you referred to
> in your last message, where can I find it?
>
