+++ environment.yml
@@ -43,4 +43,6 @@ dependencies:
     - typer==0.6.1
     - typing-extensions==4.5.0
     - networkx==3.1
-    - psutil==5.9.5    
+    - psutil==5.9.5  
+    - git+https://github.com/vikramr2/python-mincut
+    - .  
