### 2024-01-01 a260da5
add infomap to env
