### 2024-01-02 #6073 @chackoge
@LuisScoccola We looked into your issue. Vikram (the first author) has responded to it on Github. It turned out that std::binary_function is deprecated and this resulted in installation errors. python-mincut has been modified to use newer submodules and an alternate strategy (downgrading to gcc@13) is also described. I hope this addresses the problem.
