### 2024-01-04 #6073 @chryswoods
Hi - Thanks for making the changes I requested. The website with quick start and user guide is very helpful, as are the examples. I am happy that this is ready for publication.

My only comment, which is minor, is that the example relies on running an R analysis script (one of the five R scripts that are in the scripts directory). I didn't have R installed in my test environment, so this exited with an error.

Could you add "R" and any required R packages to the list of dependencies? Or, if this is only needed for some jobs, mention on the examples page that R needs to be installed to be able to run the examples.
