+++ requirements.txt
@@ -1,3 +1,4 @@
+Cython
 attrs==22.2
 click==8.1
 colorama==0.4
@@ -9,16 +10,15 @@ humanfriendly==10.0
 igraph==0.10
 iniconfig==2.0
 jsonpickle==2.2
-leidenalg==0.9
-networkit==10.1
-numpy==1.24
+leidenalg
+numpy
 packaging==23.0
-pandas==1.5
+pandas
 pip
 pluggy==1.0
 pytest==7.2
 pytz==2023.2
-scipy==1.10
+scipy
 setuptools==50.3
 six==1.16
 structlog==22.3
@@ -29,5 +29,6 @@ typing-extensions==4.5
 networkx==3.1
 psutil==5.9
 infomap==2.7
+networkit
 git+https://github.com/vikramr2/python-mincut
 .
