### 2024-01-06 #6073 @chackoge
@chryswoods We have added R as well as the data.table and feather packages as requirements in both the main readme and the documentation page. Please view the links to the updated requirements below:
ReadMe: https://github.com/illinois-or-research-analytics/cm_pipeline/tree/main?tab=readme-ov-file#requirements
Doc: https://illinois-or-research-analytics.github.io/cm_pipeline/installation/
