### 2024-01-11 #6073 @LuisScoccola
@chackoge, here are my main comments about the manuscript. I am not asking for everything to be phrased as I suggest; but if you decide to not address one of the comments in the paper, please include an explanation in a response in this github issue.

- Lines 14 and 29: Is there a precise definition of what it means for a community to be "well-connected"? Or are there many competing definitions, or is an informal notion? It would be good to add a few details about the notion of well-connectedness since it is a central motivation for your work.
- Line 29: "several widely-used codes". Are these algorithms or software libraries, or both? It would be good to clarify this in any of these cases.
- Line 30: When you cite Park et al., it would be good to be more precise and cite a more specific section of the paper, or perhaps give a few more details regarding the how this is demonstrated in that paper.
- Lines 33 and 55: Rather than saying "Key features", it would be nicer to have full sentences. Even if at the end of the day the sentences don't add any new information, I think they would improve readability. I recommend this sentences to briefly recall what "CM++" and what "CM++ Pipeline" are (even if this has been done in the introduction), so that a reader that just jumps to those sections has enough information to read them without having to go back to the introduction.
- Line 34: After "Flexibility" a colon is missing (for consistency).
- Line 50: Should "sets" be "set"?
- Line 69: I think the subsection "Parallel Strategy" is still part of the "Limitations" section. In that case (for consistency), it would be nice to either have each limitation in its own subsection, or perhaps even better, have limitations itemized in a bullet list (as the features in the previous sections), with a bold text serving as title, and then an explanation.
