+++ manuscript/paper.bib
@@ -370,3 +370,16 @@
     howpublished = {\url{https://github.com/illinois-or-research-analytics/cm_pipeline}},
     year={2023},
 }
+
+@article{doi:10.1137/040608635,
+    author = {Van Dongen, Stijn},
+    title = {Graph Clustering Via a Discrete Uncoupling Process},
+    journal = {SIAM Journal on Matrix Analysis and Applications},
+    volume = {30},
+    number = {1},
+    pages = {121-141},
+    year = {2008},
+    doi = {10.1137/040608635},
+    URL = {https://doi.org/10.1137/040608635},
+    eprint = {https://doi.org/10.1137/040608635}
+}
