+++ manuscript/paper.md
@@ -46,7 +46,7 @@ bibliography: paper.bib
 
 # Introduction
 
-Community detection methods help uncover the meso-scale structure of networks and have broad applications [@https://doi.org/10.1002/asi.22748; @Haggerty2013; @Karatas2018; @https://doi.org/10.1002/wics.1566]. While communities can be defined in different ways [@https://doi.org/10.1002/sam.10133], a common expectation is one of greater edge-density within and lesser edge density between communities [@Fortunato2022]. A related expectation is that communities also should be well-connected [@Traag2019; @BONCHI202134].
+Community detection methods help uncover the meso-scale structure of networks and have broad applications [@https://doi.org/10.1002/asi.22748; @Haggerty2013; @Karatas2018; @https://doi.org/10.1002/wics.1566]. While communities can be defined in different ways [@https://doi.org/10.1002/sam.10133], a common expectation is one of greater edge-density within and lesser edge density between communities [@Fortunato2022]. A related expectation is that communities also should be well-connected [@Traag2019; @BONCHI202134]. Well-connected clusters do not have a small minimum edge cut. Applying a mild standard for well-connected, we have previously demonstrated that several widely used implementations of community finding algorithms do not generate well-connected clusters (Park et al., 2023). A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available.
 
 In @park2023wellconnected, we describe Connectivity Modifier (CM), a meta-method that enforces well-connectedness in communities. As input, CM takes a network, a clustering of the network generated by an algorithm, and a user-specified connectivity threshold. For each community (cluster), CM uses VieCut [@henzinger2018practical] to find a small edge cut, and if the edge cut size is below the specified connectivity threshold, then the cut is removed, and the partitions are reclustered using the clustering algorithm. This process repeats until all clusters are well connected.
 
@@ -54,7 +54,7 @@ We now present CM++ [@cm_pipe2023], which uses parallelism for scalability and h
 
 # Statement of Need
 
-We have demonstrated that several widely-used algorithms and packages, such as `leidenalg` and Markov Clustering, do not generate well-connected clusters [@park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available. <!---CM++ enables a user to specify a minimum level of well-connectedness and apply it to multiple clustering paradigms. CM++ and its pipeline allow a user to choose from clustering paradigms and the level of connectedness.-->
+We have demonstrated that several widely-used algorithms and packages, such as `leidenalg` and `ikc`, do not generate well-connected clusters [see Figure 2 of @park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available. <!---CM++ enables a user to specify a minimum level of well-connectedness and apply it to multiple clustering paradigms. CM++ and its pipeline allow a user to choose from clustering paradigms and the level of connectedness.-->
 
 ![(Top) A visualization of a workflow created from the CM++ Pipeline. (Bottom right) Algorithmic schematic of CM++. CM++ splits the queue of clusters evenly between the spawned processes. Each process runs an instance of CM++ on its share of clusters (recursive mincutting and reclustering until well connected). (Bottom left) Runtime curve with respect to the number of parallel cores running CM++. CEN (14 million nodes, 1.3 billion edges) is the Curated Exosome Network. orkut (3.1 million nodes, 117 million edges) and cit_patents (3.7 million nodes, 16 million edges) are both from the SNAP database [@snapnets] and processed through the removal of parallel edges and self-loops [@park2023wellconnected].](figures/Slide3.png)
 
[…]
