+++ manuscript/paper.md
@@ -46,7 +46,7 @@ bibliography: paper.bib
 
 # Introduction
 
-Community detection methods help uncover the meso-scale structure of networks and have broad applications [@https://doi.org/10.1002/asi.22748; @Haggerty2013; @Karatas2018; @https://doi.org/10.1002/wics.1566]. While communities can be defined in different ways [@https://doi.org/10.1002/sam.10133], a common expectation is one of greater edge-density within and lesser edge density between communities [@Fortunato2022]. A related expectation is that communities also should be well-connected [@Traag2019; @BONCHI202134], which is defined by the size of the minimum edge cut. Applying a mild standard for well-connected, we have previously demonstrated that several widely used implementations of community finding algorithms do not generate well-connected clusters [@park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available.
+Community detection methods help uncover the meso-scale structure of networks and have broad applications [@https://doi.org/10.1002/asi.22748; @Haggerty2013; @Karatas2018; @https://doi.org/10.1002/wics.1566]. While communities can be defined in different ways [@https://doi.org/10.1002/sam.10133], a common expectation is one of greater edge-density within and lesser edge density between communities [@Fortunato2022]. A related expectation is that communities also should be well-connected [@Traag2019; @BONCHI202134], which is defined by the size of the minimum edge cut. Applying a mild standard for well-connectedness, we have previously demonstrated that several implementations of community finding algorithms do not generate well-connected clusters [@park2023wellconnected].
 
 In @park2023wellconnected, we describe Connectivity Modifier (CM), a meta-method that enforces well-connectedness in communities. As input, CM takes a network, a clustering of the network generated by an algorithm, and a user-specified connectivity threshold. For each community (cluster), CM uses VieCut [@henzinger2018practical] to find a small edge cut, and if the edge cut size is below the specified connectivity threshold, then the cut is removed, and the partitions are reclustered using the clustering algorithm. This process repeats until all clusters are well connected.
 
@@ -54,7 +54,7 @@ We now present CM++ [@cm_pipe2023], which uses parallelism for scalability and h
 
 # Statement of Need
 
-Previously, we have demonstrated that several implementations of community finding algorithms, e.g., MCL [@doi:10.1137/040608635], Infomap [@doi:10.1073/pnas.0706851105], Leiden [@Traag2019], IKC [@10.1162/qss_a_00184] do not generate well-connected clusters to varying extents [see Figure 2 of @park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available. <!---CM++ enables a user to specify a minimum level of well-connectedness and apply it to multiple clustering paradigms. CM++ and its pipeline allow a user to choose from clustering paradigms and the level of connectedness.-->
+Previously, we have demonstrated that several implementations of community finding algorithms, e.g., MCL [@doi:10.1137/040608635], Infomap [@doi:10.1073/pnas.0706851105], Leiden [@Traag2019] and IKC [@10.1162/qss_a_00184] generate, to varying extents, clusters that fail a mild condition for well-connectedness [see Figure 2 of @park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available. <!---CM++ enables a user to specify a minimum level of well-connectedness and apply it to multiple clustering paradigms. CM++ and its pipeline allow a user to choose from clustering paradigms and the level of connectedness.-->
 
[…]
