### 2024-01-12 #6073 @chackoge
@LuisScoccola <https://github.com/LuisScoccola>  Thank you for these
helpful comments. We have modified the manuscript as follows.
>  Lines 14 and 29: Is there a precise definition of what it means for a
community to be "well-connected"? Or are there many competing definitions,
or is an informal notion? It would be good to add a few details about the
notion of well-connectedness since it is a central motivation for your work.

The phrase “well-connected” as we and others have used (Traag, Bonchi,,
Park cited in manuscript) has to do with the size of the minimum edge cut
for the given cluster, and whether it is large enough to qualify for being
“well connected”.  The only variation in this definition has to do with the
standard imposed for what is “large enough”. We allow this to be specified
by the user,  but explore (in our study) the requirement that the min cut
size be at least log_10(n), a mild standard to impose. To us it seems that
a lengthy discussion of what well-connected really means is out of scope
and provides a level of detail that seems at variance with the purpose of
the manuscript. Instead, we've added some clarifying text indicating that
well-connectedness is defined by the mincut of a cluster/community and that
the software allows users to set a discriminating value. Accordingly...

"A related expectation is that communities also should be
well-connected (Bonchi
et al., 2021; Traag et al., 2019), which is defined by the size of the
minimum edge cut."

As mentioned, we've cited three papers in the manuscript; Bonchi et al.
(2021), Traag et al. (2019), and Park et al. (2023),  for anyone who really
wants to delve into this issue. Given the lack of a  definition that is
universally accepted, the CM++ package was enginered to allow users to
specify thresholds for well-connectedness.

> Line 29: "several widely-used codes". Are these algorithms or software
libraries, or both? It would be good to clarify this in any of these cases.

This has been addressed by specifying that they are software
implementations of algorithms with citations.

> Line 30: When you cite Park et al., it would be good to be more precise
and cite a more specific section of the paper, or perhaps give a few more
details regarding the how this is demonstrated in that paper

Yes, we have drawn the readers attention to Fig 2 of Park et al.

> Lines 33 and 55: Rather than saying "Key features", it would be nicer to
have full sentences. Even if at the end of the day the sentences don't add
any new information, I think they would improve readability. I recommend
this sentences to briefly recall what "CM++" and what "CM++ Pipeline" are
(even if this has been done in the introduction), so that a reader that
just jumps to those sections has enough information to read them without
having to go back to the introduction.

We have changed the text in response.

> Line 34: After "Flexibility" a colon is missing (for consistency).

Fixed

> Line 50: Should "sets" be "set"?

Fixed, thanks for alerting us.

> Line 69: I think the subsection "Parallel Strategy" is still part of the
"Limitations" section. In that case (for consistency), it would be nice to
either have each limitation in its own subsection, or perhaps even better,
have limitations itemized in a bullet list (as the features in the previous
sections), with a bold text serving as title, and then an explanation.

The header for that subsection has been removed and the text merged into
the Limitations section.

Reply to this email directly, view it on GitHub, or unsubscribe.
> You are receiving this because you were mentioned.
> Message ID: ***@***.***>
>
