### 2024-01-12 #6073 @LuisScoccola
@chackoge thank you for the prompt response and for addressing my comments. I have checked the paper and it looks good to me. I have three more minor comments regarding the last set of changes.

- Line 15: Regarding "Applying a mild standard for well-connectedness". I think what is meant is clear, but I don't know if I have seen the word "mild" as a modifier for "standard". If this is usual, please disregard this comment. Otherwise, please consider replacing it with something like "Using a standard  notion of well-connectedness".
- Line 34: Missing comma after "(Traag et al., 2019)".
- Line 34: Maybe say "clusters that fail to satisfy a mild condition [...]" (that is, add the words "to satisfy").
