### 2024-01-12 #6073 @chackoge
Hello @LuisScoccola:

Thanks for the quick turnaround on our revisions and the detailed comments.
We'll fix Line 34 as suggested right away.

WIth respect to Line 15 (and this goes back to my earlier comment about
scope and delving into the semantics of well connected). Mild is not used
as a modifier for standard. We used a mild definition of well-connected
defined by a slow growing function (as opposed to a stringent definition of
 well connected) when evaluating cluster quality from MCL, Infomap, Leiden
(modularity or CPM), and IKC.  Had a stringent definition been applied then
very few clusters would have satisfied the condition imposed and our
observation would have been artificially exaggerated. Hence we used 'mild'
to illustrate the prevalence of poorly connected clusters and developed
CM++ to remediate them. The function is plotted in Fig 9 of Park et al
(arXiv).

Here are relevant extracts from the text of Park that may also help.

*Extract 1.*
"We also observed that the number of clusters with small min cuts increases
as the resolution parameter decreases. Intrigued by this observation, we
performed a broader study to evaluate the extent to which clusters produced
by algorithms of interest meet even a mild standard for a well connected
cluster."

*Extract 2.*
"While a large fraction of small min cuts intuitively signals
“poorly-connected” clusters, there are different ways of formalizing this
notion. Here we offer a formal definition for the purposes of this study.
Briefly, we consider functions f(n) with the interpretation that if a
cluster of size n has an edge cut of size at most f(n) then the cluster
will be considered poorly connected. We want f(n) to grow very slowly so
that it serves as a mild bound. We also want f(n) ≥ 1 for all n that are
large enough for the cluster to be considered a potential community. From
three examples of slowly growing functions (Materials and Methods), we
choose f(n) = log10 n, the function that imposes the mildest constraint on
large clusters and grows more slowly than the bound in (10)."

Regards

George Chacko

> Line 15: Regarding "Applying a mild standard for well-connectedness". I
think what is meant is clear, but I don't know if I have seen the word
"mild" as a modifier for "standard". If this is usual, please disregard
this comment. Otherwise, please consider replacing it with something like
"Using a standard notion of well-connectedness".

Message ID: ***@***.***>
>
