+++ manuscript/paper.md
@@ -54,7 +54,7 @@ We now present CM++ [@cm_pipe2023], which uses parallelism for scalability and h
 
 # Statement of Need
 
-Previously, we have demonstrated that several implementations of community finding algorithms, e.g., MCL [@doi:10.1137/040608635], Infomap [@doi:10.1073/pnas.0706851105], Leiden [@Traag2019] and IKC [@10.1162/qss_a_00184] generate, to varying extents, clusters that fail a mild condition for well-connectedness [see Figure 2 of @park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available. <!---CM++ enables a user to specify a minimum level of well-connectedness and apply it to multiple clustering paradigms. CM++ and its pipeline allow a user to choose from clustering paradigms and the level of connectedness.-->
+Previously, we have demonstrated that several implementations of community finding algorithms, e.g., MCL [@doi:10.1137/040608635], Infomap [@doi:10.1073/pnas.0706851105], Leiden [@Traag2019], and IKC [@10.1162/qss_a_00184] generate, to varying extents, clusters that fail to satisfy a mild condition for well-connectedness [see Figure 2 of @park2023wellconnected]. A tool to enforce user-specified levels of well-connectedness to clusterings from multiple community detection methods is not presently available. <!---CM++ enables a user to specify a minimum level of well-connectedness and apply it to multiple clustering paradigms. CM++ and its pipeline allow a user to choose from clustering paradigms and the level of connectedness.-->
 
 ![(Top) A visualization of a workflow created from the CM++ Pipeline. (Bottom right) Algorithmic schematic of CM++. CM++ splits the queue of clusters evenly between the spawned processes. Each process runs an instance of CM++ on its share of clusters (recursive mincutting and reclustering until well connected). (Bottom left) Runtime curve with respect to the number of parallel cores running CM++. CEN (14 million nodes, 1.3 billion edges) is the Curated Exosome Network. orkut (3.1 million nodes, 117 million edges) and cit_patents (3.7 million nodes, 16 million edges) are both from the SNAP database [@snapnets] and processed through the removal of parallel edges and self-loops [@park2023wellconnected].](figures/Slide3.png)
 
