### 2024-01-13 #6073 @vikramr2
Hello @arfon,

I have updated the release to v4.0.1 and the Zenodo DOI is [10.5281/zenodo.10501118](https://doi.org/10.5281/zenodo.10501118).

Thanks!

Vikram
