+++ manuscript/paper.bib
@@ -364,11 +364,15 @@
 }
 
 
-@misc{cm_pipe2023,
-    author = {Vikram Ramavarapu and Vidya Kamath and Minhyuk Park and Fabio Ayres and George Chacko},
-    title = {Connectivity Modifier Pipeline},
-    howpublished = {\url{https://github.com/illinois-or-research-analytics/cm_pipeline}},
-    year={2023},
+@software{vikram_ramavarapu_2024_10501118,
+  author={Vikram Ramavarapu and Fabio Jose Ayres and Minhyuk Park and Vidya Kamath P and João Alfredo Cardoso Lamy and Tandy Warnow and George Chacko},
+  title={{CM++ - A High Performance Pipeline for Well-Connected Community Detection in Large Networks}},
+  month=jan,
+  year=2024,
+  publisher={Zenodo},
+  version={v4.0.1},
+  doi={10.5281/zenodo.10501118},
+  url={https://doi.org/10.5281/zenodo.10501118}
 }
 
 @article{doi:10.1137/040608635,
