+++ manuscript/paper.md
@@ -50,7 +50,7 @@ Community detection methods help uncover the meso-scale structure of networks an
 
 In @park2023wellconnected, we describe Connectivity Modifier (CM), a meta-method that enforces well-connectedness in communities. As input, CM takes a network, a clustering of the network generated by an algorithm, and a user-specified connectivity threshold. For each community (cluster), CM uses VieCut [@henzinger2018practical] to find a small edge cut, and if the edge cut size is below the specified connectivity threshold, then the cut is removed, and the partitions are reclustered using the clustering algorithm. This process repeats until all clusters are well connected.
 
-We now present CM++ [@cm_pipe2023], which uses parallelism for scalability and has new features. CM++  provides support for additional clustering paradigms and is designed to be extensible by other developers. Concurrently, we present the CM++ Pipeline, a modular and extensible workflow that automates CM++ operations. The pipeline consists of clustering, pre-processing, connectivity modifier (CM++), and post-processing stages with generation of cluster statistics.
+We now present CM++ [@vikram_ramavarapu_2024_10501118], which uses parallelism for scalability and has new features. CM++  provides support for additional clustering paradigms and is designed to be extensible by other developers. Concurrently, we present the CM++ Pipeline, a modular and extensible workflow that automates CM++ operations. The pipeline consists of clustering, pre-processing, connectivity modifier (CM++), and post-processing stages with generation of cluster statistics.
 
 # Statement of Need
 
