+++ manuscript/paper.bib
@@ -50,7 +50,7 @@
     eprint={2303.02813},
     archivePrefix={arXiv},
     primaryClass={cs.SI},
-    booktitle={To appear, Proceedings of COMPLEX Networks 2023}
+    booktitle={Proceedings of COMPLEX Networks 2023}
 }
 
 @article{Fortunato2022,
