### 2024-01-13 #6073 @chackoge
@arfon and we've reviewed the manuscript one last time.

Thanks

George Chacko

On Sat, Jan 13, 2024 at 2:01 PM Vikram Ramavarapu ***@***.***>
wrote:

> Hello @arfon <https://github.com/arfon>,
>
> I have updated the release to v4.0.1 and the Zenodo DOI is
> 10.5281/zenodo.10501118 <https://doi.org/10.5281/zenodo.10501118>.
>
> Thanks!
>
> Vikram
>
