+++ manuscript/paper.bib
@@ -366,7 +366,7 @@
 
 @software{vikram_ramavarapu_2024_10501118,
   author={Vikram Ramavarapu and Fabio Jose Ayres and Minhyuk Park and Vidya Kamath P and João Alfredo Cardoso Lamy and Tandy Warnow and George Chacko},
-  title={{CM++ - A High Performance Pipeline for Well-Connected Community Detection in Large Networks}},
+  title={{CM++ - A Meta-method for Well-Connected Community Detection}},
   month=jan,
   year=2024,
   publisher={Zenodo},
