### 2024-01-19 #6073 @chackoge
Well this is the best news I've had in 2024! Thank you very
much @arfon @luisscoccola @chryswoods. The article was clearly improved by
your input so we're very grateful. The journal review process was also
quite refreshing in comparison with umm..another experience we had in 2023.
Regards GC

On Fri, Jan 19, 2024 at 4:20 AM The Open Journals editorial robot <
***@***.***> wrote:

> 🎉🎉🎉 Congratulations on your paper acceptance! 🎉🎉🎉
>
> If you would like to include a link to your paper from your README use the
> following code snippets:
>
> Markdown:
>
> HTML:
> <a style="border-width:0" href="https://doi.org/10.21105/joss.06073">
> </a>
>
> reStructuredText:
>    :target: https://doi.org/10.21105/joss.06073
>
> This is how it will look in your documentation:
>
> [image: DOI] <https://doi.org/10.21105/joss.06073>
>
> *We need your help!*
>
> The Journal of Open Source Software is a community-run journal and relies
> upon volunteer effort. If you'd like to support us please consider doing
> either one (or both) of the the following:
>
>    - Volunteering to review for us sometime in the future. You can add
>    your name to the reviewer list here:
>    https://reviewers.joss.theoj.org/join
>    - Making a small donation to support our running costs here:
>    https://numfocus.org/donate-to-joss
>
